dont got a lot of time, did a lot of stuff
Umm, I renamed this file to ELlis's notes to self. I added more tiles to the "1st Leyer" tile palette, I'm thinking of deleting the "2nd Layer" tile palette. I was able to make another tile map, I think I'll use it as the groundwork for the elevation system; like stairs, testing out possible ceillings, and figuring out the collisons between the layers and the player. I probably have to make the models for a lot of stuff in order to make and use our own asset textures for them. If we used ones from the asset store assuming we found ones that fit our puporses that are also free, they'll probally look really out of place. Making the walls will be a good start to relearning how to use probuilder. Relearning it is faster than learning a new program.